Enable test_rp_raptor_staging
Resolves #119

* Allow git ref to be available when pip installing RCT.
* Allow more time for test.
* Also test local.localhost
* Don't rely on shell expansion in resource definition.
  Use explicit path to virtenv.
* Store RP logs as a GitHub Actions artifact.
  Ref: https://github.com/actions/upload-artifact
* Add mpiexec to CI image.
* Separate stable and HEAD RCT stack.
* Separate pytest tests.
  This should improve readability and maintainability,
  but is also likely necessary to get automated tests
  to pass without the file descriptor errors we have
  been seeing.
* Install missing packages when not using requirements-testing.txt
* Update the codecov config. We have 5 builds now.
