+++ .codecov.yml
@@ -2,11 +2,11 @@ codecov:
   max_report_age: off
   require_ci_to_pass: false
   notify:
-    after_n_builds: 3
-# do not notify or comment until at least 3 builds have been uploaded from the
+    after_n_builds: 5
+# do not notify or comment until all of the builds have been uploaded from the
 # CI pipeline.
 comment:
-  after_n_builds: 3
+  after_n_builds: 5
 
 fixes:
   - "*/site-packages/scalems/::src/scalems/"
