+++ requirements-testing.txt
@@ -21,6 +21,6 @@ tox>=2.0
 # radical.utils==1.6.6
 # radical.pilot==1.6.6
 # https://github.com/radical-cybertools/radical.pilot/archive/53e85ea99345b6f50f3adfd975385be3200b750f.tar.gz
-git+https://github.com/radical-cybertools/radical.utils.git@project/scalems#egg=radical.utils
-git+https://github.com/radical-cybertools/radical.saga.git@project/scalems#egg=radical.saga
-git+https://github.com/radical-cybertools/radical.pilot.git@project/scalems#egg=radical.pilot
+git+https://github.com/radical-cybertools/radical.utils.git@scalems/stable#egg=radical.utils
+git+https://github.com/radical-cybertools/radical.saga.git@scalems/stable#egg=radical.saga
+git+https://github.com/radical-cybertools/radical.pilot.git@scalems/stable#egg=radical.pilot
